refactor(dataobj): move dataobj encoding to dataobj package
