+++ pkg/dataobj/decoder.go
@@ -1,0 +1,67 @@
+package dataobj
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/grafana/loki/v3/pkg/dataobj/internal/metadata/filemd"
+	"github.com/grafana/loki/v3/pkg/dataobj/internal/util/bufpool"
+)
+
+type decoder struct {
+	rr rangeReader
+}
+
+func (d *decoder) Metadata(ctx context.Context) (*filemd.Metadata, error) {
+	tailer, err := d.tailer(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("reading tailer: %w", err)
+	}
+
+	rc, err := d.rr.ReadRange(ctx, int64(tailer.FileSize-tailer.MetadataSize-8), int64(tailer.MetadataSize))
+	if err != nil {
+		return nil, fmt.Errorf("getting metadata: %w", err)
+	}
+	defer rc.Close()
+
+	br := bufpool.GetReader(rc)
+	defer bufpool.PutReader(br)
+
+	return decodeFileMetadata(br)
+}
+
+type tailer struct {
+	MetadataSize uint64
+	FileSize     uint64
+}
+
+func (d *decoder) tailer(ctx context.Context) (tailer, error) {
+	size, err := d.rr.Size(ctx)
+	if err != nil {
+		return tailer{}, fmt.Errorf("reading attributes: %w", err)
+	}
+
+	// Read the last 8 bytes of the object to get the metadata size and magic.
+	rc, err := d.rr.ReadRange(ctx, size-8, 8)
+	if err != nil {
+		return tailer{}, fmt.Errorf("getting file tailer: %w", err)
+	}
+	defer rc.Close()
+
+	br := bufpool.GetReader(rc)
+	defer bufpool.PutReader(br)
+
+	metadataSize, err := decodeTailer(br)
+	if err != nil {
+		return tailer{}, fmt.Errorf("scanning tailer: %w", err)
+	}
+
+	return tailer{
+		MetadataSize: uint64(metadataSize),
+		FileSize:     uint64(size),
+	}, nil
+}
+
+func (d *decoder) SectionReader(metadata *filemd.Metadata, section *filemd.SectionInfo) SectionReader {
+	return &sectionReader{rr: d.rr, md: metadata, sec: section}
+}
