+++ pkg/dataobj/section_reader.go
@@ -1,4 +1,4 @@
-package encoding
+package dataobj
 
 import (
 	"context"
@@ -7,88 +7,15 @@ import (
 	"math"
 
 	"github.com/grafana/loki/v3/pkg/dataobj/internal/metadata/filemd"
-	"github.com/grafana/loki/v3/pkg/dataobj/internal/util/bufpool"
 )
 
-// rangeReader is an interface that can read a range of bytes from an object.
-type rangeReader interface {
-	// Size returns the full size of the object.
-	Size(ctx context.Context) (int64, error)
-
-	// ReadRange returns a reader over a range of bytes. Callers may create
-	// multiple current instance of ReadRange.
-	ReadRange(ctx context.Context, offset int64, length int64) (io.ReadCloser, error)
-}
-
-type rangeDecoder struct {
-	r rangeReader
-}
-
-func (rd *rangeDecoder) Metadata(ctx context.Context) (*filemd.Metadata, error) {
-	tailer, err := rd.tailer(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("reading tailer: %w", err)
-	}
-
-	rc, err := rd.r.ReadRange(ctx, int64(tailer.FileSize-tailer.MetadataSize-8), int64(tailer.MetadataSize))
-	if err != nil {
-		return nil, fmt.Errorf("getting metadata: %w", err)
-	}
-	defer rc.Close()
-
-	br := bufpool.GetReader(rc)
-	defer bufpool.PutReader(br)
-
-	return decodeFileMetadata(br)
-}
-
-type tailer struct {
-	MetadataSize uint64
-	FileSize     uint64
-}
-
-func (rd *rangeDecoder) tailer(ctx context.Context) (tailer, error) {
-	size, err := rd.r.Size(ctx)
-	if err != nil {
-		return tailer{}, fmt.Errorf("reading attributes: %w", err)
-	}
-
-	// Read the last 8 bytes of the object to get the metadata size and magic.
-	rc, err := rd.r.ReadRange(ctx, size-8, 8)
-	if err != nil {
-		return tailer{}, fmt.Errorf("getting file tailer: %w", err)
-	}
-	defer rc.Close()
-
-	br := bufpool.GetReader(rc)
-	defer bufpool.PutReader(br)
-
-	metadataSize, err := decodeTailer(br)
-	if err != nil {
-		return tailer{}, fmt.Errorf("scanning tailer: %w", err)
-	}
-
-	return tailer{
-		MetadataSize: uint64(metadataSize),
-		FileSize:     uint64(size),
-	}, nil
-}
-
-func (rd *rangeDecoder) SectionReader(metadata *filemd.Metadata, section *filemd.SectionInfo) SectionReader {
-	return &rangeSectionReader{rr: rd.r, md: metadata, sec: section}
-}
-
-type rangeSectionReader struct {
+type sectionReader struct {
 	rr  rangeReader // Reader for absolute ranges within the file.
 	md  *filemd.Metadata
 	sec *filemd.SectionInfo
 }
 
-func (rd *rangeSectionReader) Type() (SectionType, error) {
-	return GetSectionType(rd.md, rd.sec)
-}
-
-func (rd *rangeSectionReader) DataRange(ctx context.Context, offset, length int64) (io.ReadCloser, error) {
+func (sr *sectionReader) DataRange(ctx context.Context, offset, length int64) (io.ReadCloser, error) {
 	if offset < 0 || length < 0 {
 		return nil, fmt.Errorf("parameters must not be negative: offset=%d length=%d", offset, length)
 	}
@@ -100,7 +27,7 @@ func (rd *rangeSectionReader) DataRange(ctx context.Context, offset, length int6
 	// offset if a layout is provided.
 	var absoluteOffset = offset
 
-	if layout := rd.sec.Layout; layout != nil {
+	if layout := sr.sec.Layout; layout != nil {
 		if layout.Data == nil {
 			return nil, fmt.Errorf("section has no data")
 		} else if layout.Data.Offset > math.MaxInt64 {
@@ -121,18 +48,18 @@ func (rd *rangeSectionReader) DataRange(ctx context.Context, offset, length int6
 		absoluteOffset = int64(layout.Data.Offset) + offset
 	}
 
-	return rd.rr.ReadRange(ctx, absoluteOffset, length)
+	return sr.rr.ReadRange(ctx, absoluteOffset, length)
 }
 
-func (rd *rangeSectionReader) Metadata(ctx context.Context) (io.ReadCloser, error) {
-	metadataRegion, err := findMetadataRegion(rd.sec)
+func (sr *sectionReader) Metadata(ctx context.Context) (io.ReadCloser, error) {
+	metadataRegion, err := findMetadataRegion(sr.sec)
 	if err != nil {
 		return nil, err
 	} else if metadataRegion == nil {
 		return nil, fmt.Errorf("section is missing metadata")
 	}
 
-	return rd.rr.ReadRange(ctx, int64(metadataRegion.Offset), int64(metadataRegion.Length))
+	return sr.rr.ReadRange(ctx, int64(metadataRegion.Offset), int64(metadataRegion.Length))
 }
 
 // findMetadataRegion returns the region where a section's metadata is stored.
